+++ src/api/transports/index.js
@@ -1,0 +1,7 @@
+import HttpTransport from './http';
+import WsTransport from './ws';
+
+export default {
+  http: HttpTransport,
+  ws: WsTransport,
+};
